+++ HISTORY.md
@@ -1,5 +1,5 @@
-unreleased
-==========
+0.6.2 / 2019-04-29
+==================
 
   * Fix sorting charset, encoding, and language with extra parameters